Refactor TitleContainsKeywordsPredicate to SourceContainsKeywordsPredicate
