--- src/main/java/seedu/address/model/source/TitleContainsKeywordsPredicate.java
@@ -1,31 +1,0 @@
-package seedu.address.model.source;
-
-import java.util.List;
-import java.util.function.Predicate;
-
-import seedu.address.commons.util.StringUtil;
-
-/**
- * Tests that a {@code Source}'s {@code Title} matches any of the keywords given.
- */
-public class TitleContainsKeywordsPredicate implements Predicate<Source> {
-    private final List<String> keywords;
-
-    public TitleContainsKeywordsPredicate(List<String> keywords) {
-        this.keywords = keywords;
-    }
-
-    @Override
-    public boolean test(Source source) {
-        return keywords.stream()
-                .anyMatch(keyword -> StringUtil.containsWordIgnoreCase(source.getTitle().title, keyword));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof TitleContainsKeywordsPredicate // instanceof handles nulls
-                && keywords.equals(((TitleContainsKeywordsPredicate) other).keywords)); // state check
-    }
-
-}
